Add an utility class for gateway certificate management
Add an utility class for gateway certificate management
Add new configuration enable_certificate_chain_validation to api-manager.xml.j2
