+++ components/apimgt/org.wso2.carbon.apimgt.gateway/src/main/java/org/wso2/carbon/apimgt/gateway/handlers/security/authenticator/MutualSSLAuthenticator.java
@@ -207,7 +207,7 @@ public class MutualSSLAuthenticator implements Authenticator {
             subjectDNIdentifiers.add(subjectDNIdentifier);
             for (Map.Entry<String, String> entry : certificates.entrySet()) {
                 String key = entry.getKey();
-                if (key.contains(subjectDNIdentifier)) {
+                if (StringUtils.equals(subjectDNIdentifier, key)) {
                     uniqueIdentifier = key;
                     tier = entry.getValue();
                     break;
